added response behavior in accordance with the REST methodology,headers with frameOptions for h2-console and other minor changes in product and order controllers

+++ src/main/java/com/example/springcoredemo/controller/OrderController.java
@@ -4,7 +4,10 @@ package com.example.springcoredemo.controller;
 import com.example.springcoredemo.model.OrderDTO;
 import com.example.springcoredemo.service.OrderService;
 import com.example.springcoredemo.utils.Util;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import com.example.springcoredemo.repository.OrderRepository;
 
 import java.util.List;
 
@@ -14,9 +17,11 @@ import java.util.List;
 public class OrderController {
 
     private final OrderService orderService;
+    private final OrderRepository orderRepository;
 
-    public OrderController(OrderService orderService) {
+    public OrderController(OrderService orderService, OrderRepository orderRepository) {
         this.orderService = orderService;
+        this.orderRepository = orderRepository;
     }
 
     @GetMapping("/{id}")
@@ -30,22 +35,46 @@ public class OrderController {
     }
 
     @PostMapping
-    public void save(@RequestBody OrderDTO orderDTO) {
-        if (Util.anyNull(orderDTO.getDate()))
+    public ResponseEntity<?> save(@RequestBody OrderDTO orderDTO) {
+        if (Util.anyNull(orderDTO.getDate())) {
             throw new IllegalArgumentException("Can't save order without date");
-        orderService.save(orderDTO);
+        }
+
+        Integer orderId = orderDTO.getId();
+
+
+        if (orderId != null && orderRepository.existsById(orderId)) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Order with id " + orderId + " already exists");
+        }
+
+        OrderDTO savedOrderDTO = orderService.save(orderDTO);
+
+
+        return ResponseEntity.status(HttpStatus.CREATED).body(savedOrderDTO);
     }
+
 
     @PutMapping
-    public void update(@RequestBody OrderDTO orderDTO) {
-        if (Util.anyNull(orderDTO.getId()))
+    public ResponseEntity<OrderDTO> update(@RequestBody OrderDTO orderDTO) {
+        if (orderDTO.getId() == null) {
             throw new IllegalArgumentException("Can't update order without id");
-        orderService.update(orderDTO);
+        }
+
+        OrderDTO updatedOrderDTO = orderService.update(orderDTO);
+
+        return ResponseEntity.ok(updatedOrderDTO);
     }
+
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable Integer id) {
-        orderService.delete(id);
-    }
+    public ResponseEntity<String> delete(@PathVariable Integer id) {
 
+        if (!orderRepository.existsById(id)) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Order not found with id: " + id);
+        }
+
+        orderService.delete(id);
+
+        return ResponseEntity.ok("Order deleted successfully");
+    }
 }
